Fix signedness warnings in glib/tests/sequence.c
Note that this patch slightly change the API of glib/gsequence.c from
gint g_sequence_get_length() to gsize g_sequence_get_length().

In file included from glib/glibconfig.h:9,
                 from glib/gtypes.h:32,
                 from glib/galloca.h:32,
                 from glib/glib.h:30,
                 from glib/tests/sequence.c:2:
glib/tests/sequence.c: In function ‘check_integrity’:
glib/tests/sequence.c:139:27: error: comparison of integer expressions of different signedness: ‘int’ and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
   g_assert (info->n_items == g_queue_get_length (info->queue));
                           ^~
glib/gmacros.h:900:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
glib/tests/sequence.c:139:3: note: in expansion of macro ‘g_assert’
   g_assert (info->n_items == g_queue_get_length (info->queue));
   ^~~~~~~~
glib/tests/sequence.c:157:27: error: comparison of integer expressions of different signedness: ‘int’ and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
   g_assert (info->n_items == g_queue_get_length (info->queue));
                           ^~
glib/gmacros.h:900:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
glib/tests/sequence.c:157:3: note: in expansion of macro ‘g_assert’
   g_assert (info->n_items == g_queue_get_length (info->queue));
   ^~~~~~~~
glib/tests/sequence.c: In function ‘run_random_tests’:
glib/tests/sequence.c:554:55: error: comparison of integer expressions of different signedness: ‘guint’ {aka ‘unsigned int’} and ‘gint’ {aka ‘int’} [-Werror=sign-compare]
             g_assert (g_queue_get_length (seq->queue) == g_sequence_get_length (seq->sequence));
                                                       ^~
glib/gmacros.h:900:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
glib/tests/sequence.c:554:13: note: in expansion of macro ‘g_assert’
             g_assert (g_queue_get_length (seq->queue) == g_sequence_get_length (seq->sequence));
             ^~~~~~~~
glib/tests/sequence.c: In function ‘main’:
glib/tests/sequence.c:1404:17: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘long unsigned int’ [-Werror=sign-compare]
   for (i = 0; i < G_N_ELEMENTS (seeds); ++i)
                 ^

+++ glib/gsequence.h
@@ -41,7 +41,7 @@ GSequence *    g_sequence_new                (GDestroyNotify            data_des
 GLIB_AVAILABLE_IN_ALL
 void           g_sequence_free               (GSequence                *seq);
 GLIB_AVAILABLE_IN_ALL
-gint           g_sequence_get_length         (GSequence                *seq);
+gsize          g_sequence_get_length         (GSequence                *seq);
 GLIB_AVAILABLE_IN_ALL
 void           g_sequence_foreach            (GSequence                *seq,
                                               GFunc                     func,
